examples/computer_use/requirements.txt
python-dotenv
pyautogui
openai
Pillow